Closes boozdb-4, add docker build. Used the maven execute plugin to perform docker clean and build during maven phase install.
