--- src/main/java/com/boozdb/MainConfiguration.java
@@ -1,25 +1,0 @@
-package com.boozdb;
-
-import io.dropwizard.Configuration;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.dropwizard.db.DataSourceFactory;
-
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-
-public class MainConfiguration extends Configuration {
-
-    @Valid
-    @NotNull
-    private DataSourceFactory database = new DataSourceFactory();
-
-    @JsonProperty("database")
-    public void setDataSourceFactory(DataSourceFactory factory) {
-        this.database = factory;
-    }
-
-    @JsonProperty("database")
-    public DataSourceFactory getDataSourceFactory() {
-        return database;
-    }
-}
