--- src/main/java/com/boozdb/service/views/HomeView.java
@@ -1,13 +1,0 @@
-package com.boozdb.service.views;
-
-import io.dropwizard.views.View;
-
-public class HomeView extends View {
-
-    final public BoozDbMeta meta;
-
-    public HomeView(BoozDbMeta meta) {
-        super("index.mustache");
-        this.meta = meta;
-    }
-}
