+++ src/main/java/com/boozedb/core/constants/DefaultParameterValues.java
@@ -1,0 +1,16 @@
+package com.boozedb.core.constants;
+
+public class DefaultParameterValues {
+
+    public final static double MIN_PRICE = 0;
+    public final static double MAX_PRICE = 0;
+    public final static int MIN_AGE = 0;
+    public final static int MAX_AGE = 0;
+    public final static double  MIN_PROOF = 0;
+    public final static double  MAX_PROOF = 0;
+    public final static double  MIN_SIZE = 0;
+    public final static double  MAX_SIZE = 0;
+    public final static int PAGE_OFFSET = 0;
+    public final static int  PAGE_SIZE = 20;
+
+}
