+++ src/main/java/com/boozedb/service/views/HomeView.java
@@ -1,0 +1,13 @@
+package com.boozedb.service.views;
+
+import io.dropwizard.views.View;
+
+public class HomeView extends View {
+
+    final public BoozeDbMeta meta;
+
+    public HomeView(BoozeDbMeta meta) {
+        super("index.mustache");
+        this.meta = meta;
+    }
+}
