+++ src/main/resources/banner.txt
@@ -1,6 +1,6 @@
 ================================================================================
 
-                              BoozDb
+                              BoozeDb
 
 ================================================================================
