--- src/main/resources/com/boozdb/service/views/index.mustache
@@ -1,42 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
-    <link rel="stylesheet" href="assets/style.css">
-
-    <title>BoozDb {{#bottleResponse.bottle}}{{bottleResponse.bottle.description}}{{/bottleResponse.bottle}}</title>
-
-</head>
-<body>
-<nav class="navbar navbar-expand-md navbar-dark bg-dark mb-4 sticky-top" id="nav">
-    <a class="navbar-brand mr-0 mr-md-2" href="/" aria-label="BoozDb">BoozDb</a>
-</nav>
-<div class="container" id="content">
-    {{#bottleResponse.bottle}}
-        {{> partials/bottle}}
-    {{/bottleResponse.bottle}}
-
-    {{#bottleListResponse}}
-        {{> partials/bottleList}}
-    {{/bottleListResponse}}
-
-    {{^bottleResponse.bottle}}
-        {{^bottleListResponse}}
-            {{> partials/home}}
-        {{/bottleListResponse}}
-
-    {{/bottleResponse.bottle}}
-
-</div>
-<!-- Optional JavaScript -->
-<!-- jQuery first, then Popper.js, then Bootstrap JS -->
-<script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js" integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN" crossorigin="anonymous"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js" integrity="sha384-B4gt1jrGC7Jh4AgTPSdUtOBvfO8shuf57BaghqFfPlYxofvL8/KUEfYiJOMMV+rV" crossorigin="anonymous"></script>
-</body>
-</html>
