+++ src/main/resources/com/boozedb/service/views/partials/bottle.mustache
