+++ src/main/resources/com/boozedb/service/views/partials/bottleCard.mustache
