+++ src/main/resources/com/boozedb/service/views/partials/bottleList.mustache
