+++ src/main/resources/com/boozedb/service/views/partials/home.mustache
