+++ src/main/resources/com/boozedb/service/views/partials/listItem.mustache
