+++ src/main/resources/com/boozedb/service/views/partials/search/category.mustache
