+++ src/main/resources/com/boozedb/service/views/partials/search/categoryButtons.mustache
