+++ src/main/resources/com/boozedb/service/views/partials/search/searchControls.mustache
