+++ src/main/resources/com/boozedb/service/views/partials/search/subcategoriesLi.mustache
